Drop casting of time.time() to int, during sample collection.
Currently, when collecting samples, and when legacy_timestamp has been set to True, the time.time() value is cast to int, which is reducing the precision of the timestamp. There does not seem to be any reason to return an int for timestamp. And, the documentation for the function mentions that if legacy_timestamp is set to True, it will return time.time(). So this change is more accurate given existing docstring, and will increase the precision of the timestamp value returned.